fix: patch path finding
Surprisingly it is not enough to just strip some prefix from the patch
paths, sometimes you also have to find shortest possible matching
subpath from the working directory.

We fix that, and correspondingly change test to check only if diffy
succeedes as git is broken here and there.

There are still 4 test cases where git and diffy give different
results, but that is due to git incorrect behavior. If you look at the
directory diff output and patch files you'll see that diffy applies
patches correctly and git silently fails. Comment is added with exact
recipes that fail, so look at it if you want to check it out yourself.

There are also couple of packages that still fail and we ignore
them (they include mumps substring in their name) since their format
is looking strange, but it is possible to easily fix this issue in
practice. Final decision is postponed, since further investigation is
needed on why it even works in conda.  Following is the header of one
of the patches in mumps recipe that fails.

```diff
--- mumps/src/mumps_common_orig.h
+++ mumps/src/mumps_common.h
```

In actual directory there is no file with `_orig` postfix, therefore
we fail to find it.

In total there is 6 different test cases that fail and we ignore it
due to the reasons described above.